refactor: split json backup into files
