+++ src/utils/type-transform/BackupFormat/JSON/latest.ts
@@ -1,0 +1,6 @@
+import { BackupJSONFileVersion1 } from './version-1'
+
+/**
+ * Always use this interface in other code.
+ */
+export interface BackupJSONFileLatest extends BackupJSONFileVersion1 {}
