+++ src/utils/type-transform/BackupFormat/JSON/version-2.ts
@@ -1,0 +1,1 @@
+export interface BackupJSONFileVersion2 {}
